+++ src/ethereum/shanghai/fork_types.py
@@ -16,10 +16,11 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING, Tuple, Union
 
 if TYPE_CHECKING:
-    from .transactions import LegacyTransaction, Withdrawal
+    from .transactions import LegacyTransaction
 
 from .. import rlp
 from ..base_types import (
+    U64,
     U256,
     Bytes,
     Bytes8,
@@ -71,6 +72,19 @@ def encode_account(raw_account_data: Account, storage_root: Bytes) -> Bytes:
             keccak256(raw_account_data.code),
         )
     )
+
+
+@slotted_freezable
+@dataclass
+class Withdrawal:
+    """
+    Withdrawals that have been validated on the consensus layer.
+    """
+
+    index: U64
+    validator_index: U64
+    address: Address
+    amount: U256
 
 
 @slotted_freezable
