+++ src/ethereum/shanghai/trie.py
@@ -36,8 +36,15 @@ from ethereum.utils.hexadecimal import hex_to_bytes
 
 from .. import rlp
 from ..base_types import U256, Bytes, Uint, slotted_freezable
-from .fork_types import Account, Address, Receipt, Root, encode_account
-from .transactions import LegacyTransaction, Withdrawal
+from .fork_types import (
+    Account,
+    Address,
+    Receipt,
+    Root,
+    Withdrawal,
+    encode_account,
+)
+from .transactions import LegacyTransaction
 
 # note: an empty trie (regardless of whether it is secured) has root:
 #
